+++ citelang/main/packages/pypi_requirements.py
@@ -2,37 +2,20 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-# Parse a requirements.txt file to generate a "package"
-import citelang.main.endpoints as endpoints
-
 import re
 
 from .base import PackagesFromFile
 
 
-class RequirementsManager(PackagesFromFile):
-    """
-    Packages parsed from a requirements.txt file (so from cran)
-    """
-
-    name = "requirements.txt"
-    underlying_manager = "pypi"
-    project_count = None
-    homepage = "pypi.org/"
-    color = "#006dad"
-    default_language = "Python"
-    default_versions = None
-
-    def parse(self, content):
+class PythonManager(PackagesFromFile):
+    def parse_python_deps(self, lines):
         """
-        Parse the self.content (the requirements.txt file)
+        Shared function for deriving a list of python dependencies from lines
         """
-        repo = self.get_repo()
-
         # Dependencies we parse as pypi packages
         # This should also update the cache and make it easier to retrieve later
         deps = []
-        for line in content.split("\n"):
+        for line in lines:
 
             # We can't easily add github pypi references
             if not line or "git@" in line:
@@ -48,25 +31,7 @@ class RequirementsManager(PackagesFromFile):
                 continue
 
             # First add requirements (names and pypi manager) to deps
-
-            # Try to get from cache - either versioned or not
-            pkg = None
-            if package_name and version:
-                cache_name = f"package/pypi/{package_name}/{version}"
-                result = self.cache.get(cache_name)
-                if result:
-                    pkg = endpoints.get_endpoint("package", data=result)
-
-            elif package_name and not version:
-                cache_name = f"package/pypi/{package_name}"
-                result = self.cache.get(cache_name)
-                if result:
-                    pkg = endpoints.get_endpoint("package", data=result)
-
-            if pkg is None:
-                pkg = endpoints.get_endpoint(
-                    "package", package_name=package_name, manager="pypi"
-                )
+            pkg = self.get_package(package_name, version)
 
             # Ensure we have version, fallback to latest
             if not version:
@@ -89,7 +54,86 @@ class RequirementsManager(PackagesFromFile):
                 "repository_sources": ["Pypi"],
             }
             deps.append(dep)
+        return deps
 
+
+class RequirementsManager(PythonManager):
+
+    """
+    Packages parsed from a requirements.txt file (so from cran)
+    """
+
+    name = "requirements.txt"
+    underlying_manager = "pypi"
+    project_count = None
+    homepage = "pypi.org/"
+    color = "#006dad"
+    default_language = "Python"
+    default_versions = None
+
+    def parse(self, content):
+        """
+        Parse the self.content (the requirements.txt file)
+        """
+        repo = self.get_repo()
+        lines = content.split("\n")
+        deps = self.parse_python_deps(lines)
+        repo["dependencies"] = deps
+        self.data["package"] = repo
+        self.data["dependencies"] = deps
+        return repo
+
+
+class SetupManager(PythonManager):
+
+    """
+    Packages parsed from a requirements.txt file (so from cran)
+    """
+
+    name = "setup.py"
+    underlying_manager = "pypi"
+    project_count = None
+    homepage = "pypi.org/"
+    color = "#006dad"
+    default_language = "Python"
+    default_versions = None
+
+    def parse(self, content):
+        """
+        Parse the self.content (the requirements.txt file)
+        """
+        repo = self.get_repo()
+
+        lines = []
+        parsing = False
+        for line in content.split("\n"):
+
+            if "setup_requires" in line:
+                parsing = True
+
+            # We found the start and end
+            if parsing and "[" in line and "]" in line:
+                lines.append(line)
+                break
+            elif parsing and "[" in line:
+                lines.append(line)
+
+            elif parsing and "]" in line:
+                lines.append(line)
+                break
+            elif parsing:
+                lines.append(line)
+
+        # Clean up lines
+        cleaned = []
+        for line in lines:
+            terms = ["setup_requires" + x for x in [":", " :", "=", " =", ""]]
+            for term in terms + ["[", "]", '"', "'"]:
+                line = line.replace(term, "")
+            parts = line.split(",")
+            cleaned += [x.strip() for x in parts if x.strip()]
+
+        deps = self.parse_python_deps(cleaned)
         repo["dependencies"] = deps
         self.data["package"] = repo
         self.data["dependencies"] = deps